refactor: references to calender
This is a typo, and should instead be calendar. Various files and strings require correction, as well as entire db schema
