+++ .cspell.json
@@ -18,5 +18,5 @@
     "yarn.lock"
   ],
   "words": ["cowpeas", "Chipata", "CLIMWAT", "Cropwat", "e-picsa", "Nkhotakota", "picsa", "sheller", "terciles"],
-  "flagWords": []
+  "flagWords": ["calender"]
 }
