+++ apps/picsa-tools/seasonal-calendar-tool/src/app/components/calendar-editor/calendar-editor.component.html
@@ -18,7 +18,7 @@
   <!-- Months generators -->
   <div>
     <h2>{{ 'Calendar Time' | translate }}</h2>
-    <div class="calender-box">
+    <div class="calendar-box">
       <mat-form-field>
         <mat-label>{{ 'Start Month ' | translate }} :</mat-label>
         <mat-select id="startMonth" [formControl]="monthForm.controls.startIndex">
