+++ apps/picsa-tools/seasonal-calendar-tool/src/app/components/calendar-editor/calendar-editor.component.scss
@@ -10,7 +10,7 @@
     height: 40px;
   }
 }
-.calender-box {
+.calendar-box {
   display: flex;
   gap: 2rem;
 }
