+++ apps/picsa-tools/seasonal-calendar-tool/src/app/pages/calendar-table/calendar-table.component.scss
