+++ apps/picsa-tools/seasonal-calendar-tool/src/app/pages/calendar-table/calendar-table.module.ts
@@ -26,4 +26,4 @@ const routes: Routes = [
     SeasonalCalendarToolComponentsModule,
   ],
 })
-export class CalenderTableModule {}
+export class CalendarTableModule {}
