+++ apps/picsa-tools/seasonal-calendar-tool/src/app/pages/home/home.component.html
@@ -3,8 +3,8 @@
   <div class="calendar-list">
     <div class="feedbackMessage" *ngIf="dbCalendars.length === 0">{{ 'No saved calendars' | translate }}</div>
     <!-- Calendar list items -->
-    @for (calendar of dbCalendars; track calendar.ID) {
-    <mat-card class="calendars-display" [routerLink]="calendar.ID">
+    @for (calendar of dbCalendars; track calendar.id) {
+    <mat-card class="calendars-display" [routerLink]="calendar.id">
       <mat-card-content>
         <div style="display: flex">
           <mat-card-title style="flex: 1">
@@ -20,7 +20,7 @@
   </div>
   <!-- Create calendar button -->
   <div class="sticky-footer">
-    <button class="footer-button" routerLink="create" mat-raised-button color="primary">
+    <button routerLink="create" mat-raised-button color="primary">
       <mat-icon>add</mat-icon>
       {{ 'Create Calendar' | translate }}
     </button>
