+++ apps/picsa-tools/seasonal-calendar-tool/src/app/pages/home/home.routing.module.ts
@@ -1,23 +1,23 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { CalendarTableComponent } from '../calender-table/calendar-table.component';
+import { CalendarTableComponent } from '../calendar-table/calendar-table.component';
 import { CreateCalendarComponent } from '../create-calendar/create-calendar.component';
 import { HomeComponent } from './home.component';
 
 const routes: Routes = [
-    {
-        path: '',
-        component: HomeComponent,
-    },
-    {
-      path: 'create',
-      component: CreateCalendarComponent,
-     },
-     {
-      path: ':id',
-      component: CalendarTableComponent,
-     }
+  {
+    path: '',
+    component: HomeComponent,
+  },
+  {
+    path: 'create',
+    component: CreateCalendarComponent,
+  },
+  {
+    path: ':id',
+    component: CalendarTableComponent,
+  },
 ];
 
 @NgModule({
