+++ libs/i18n/assets/en.json
@@ -118,8 +118,11 @@
   "Total Consumed": "Total Consumed",
   "Total Produced": "Total Produced",
   "What is your type of enterprise?": "What is your type of enterprise?",
+  "balance": "balance",
+  "consumed": "consumed",
   "description": "description",
   "fruits": "fruits",
+  "hours": "hours",
   "livestock": "livestock",
   "starting in": "starting in",
   "title": "title",
@@ -372,13 +375,13 @@
   "Title": "Title",
   "Type": "Type",
   "Calendar Time": "Calendar Time",
-  "Calender name": "Calender name",
+  "Calendar name": "Calendar name",
+  "Copy": "Copy",
   "Create Calendar": "Create Calendar",
-  "Create calender": "Create calender",
   "Example Calendar": "Example Calendar",
   "No saved calendars": "No saved calendars",
   "Number of Months": "Number of Months",
-  "Please add the calender title": "Please add the calender title",
+  "Please add title": "Please add title",
   "Please enter start month": "Please enter start month",
   "Please enter total months": "Please enter total months",
   "Please select at least 1 crop": "Please select at least 1 crop",
