+++ libs/i18n/assets/mw_ny.json
@@ -29,6 +29,7 @@
   "Bad": "•Bad•",
   "bagging": "kupakila mmatumba",
   "bags": "matumba",
+  "balance": "•balance•",
   "Balance": "Ndalama zotsala",
   "banding": "kubandila",
   "Beans": "Nyemba",
@@ -42,8 +43,8 @@
   "build housing": "kumanga makola",
   "Calculated by using FAO CLIMWAT 2.0 for Cropwat and Cropwat 8.0 and climate data for Nkhotakota Met Station.": "Kuwerengeredwa pogwiritsa ntchito FAO CLIMWAT 2.0 ya Cropwat ndi Cropwat 8.0 ndi uthenga wa  zanyengo ya Nkhotakota Met Station.",
   "Calculated with the date range of Start of rains and End of the rains": "•Calculated with the date range of Start of rains and End of the rains•",
+  "Calendar name": "•Calendar name•",
   "Calendar Time": "•Calendar Time•",
-  "Calender name": "•Calender name•",
   "Cancel": "thetsani",
   "Capacity of 100mm on 1st March reduces 0.5 with evaporation rate of 5mm per day": "•Capacity of 100mm on 1st March reduces 0.5 with evaporation rate of 5mm per day•",
   "Cassava": "Chinangwa",
@@ -62,13 +63,14 @@
   "Compare different options and plans": "Siyanitsani zisankho ndi ma pulani osiyanasiyana",
   "Comparing different options and planning using Participatory Budgets": "•Comparing different options and planning using Participatory Budgets•",
   "compost manure making": "kupanga manyowa",
+  "consumed": "•consumed•",
   "Consumed": "Zodyedwa",
+  "Copy": "•Copy•",
   "Cost": "Mtengo",
   "Cotton": "•Cotton•",
   "Cowpeas": "•Cowpeas•",
   "cowpeas": "khobwe",
   "Create Calendar": "•Create Calendar•",
-  "Create calender": "•Create calender•",
   "Create New": "•Create New•",
   "Create New Budget": "Pangani bajeti yatsopano",
   "Crop": "•Crop•",
@@ -139,6 +141,7 @@
   "Hide Options": "Bisani Zipangizo",
   "High": "•High•",
   "Historic climate": "mbiri ya za nyengo",
+  "hours": "•hours•",
   "housing": "khola",
   "How long is your budget for?": "bajeti yanu imakhala yayitali bwanji?",
   "Import": "Kutenga bajeti yokonzedwa",
@@ -229,7 +232,7 @@
   "Pigeon Peas": "•Pigeon Peas•",
   "pigeon peas": "nandolo",
   "pigs": "nkhumba",
-  "Please add the calender title": "•Please add the calender title•",
+  "Please add title": "•Please add title•",
   "Please enter start month": "•Please enter start month•",
   "Please enter total months": "•Please enter total months•",
   "Please select at least 1 crop": "•Please select at least 1 crop•",
