+++ libs/i18n/assets/sw.json
@@ -29,6 +29,7 @@
   "Bad": "•Bad•",
   "bagging": "•bagging•",
   "bags": "•bags•",
+  "balance": "•balance•",
   "Balance": "•Balance•",
   "banding": "•banding•",
   "Beans": "•Beans•",
@@ -42,8 +43,8 @@
   "build housing": "•build housing•",
   "Calculated by using FAO CLIMWAT 2.0 for Cropwat and Cropwat 8.0 and climate data for Nkhotakota Met Station.": "•Calculated by using FAO CLIMWAT 2.0 for Cropwat and Cropwat 8.0 and climate data for Nkhotakota Met Station.•",
   "Calculated with the date range of Start of rains and End of the rains": "•Calculated with the date range of Start of rains and End of the rains•",
+  "Calendar name": "•Calendar name•",
   "Calendar Time": "•Calendar Time•",
-  "Calender name": "•Calender name•",
   "Cancel": "•Cancel•",
   "Capacity of 100mm on 1st March reduces 0.5 with evaporation rate of 5mm per day": "•Capacity of 100mm on 1st March reduces 0.5 with evaporation rate of 5mm per day•",
   "Cassava": "•Cassava•",
@@ -62,13 +63,14 @@
   "Compare different options and plans": "•Compare different options and plans•",
   "Comparing different options and planning using Participatory Budgets": "•Comparing different options and planning using Participatory Budgets•",
   "compost manure making": "•compost manure making•",
+  "consumed": "•consumed•",
   "Consumed": "•Consumed•",
+  "Copy": "•Copy•",
   "Cost": "•Cost•",
   "Cotton": "•Cotton•",
   "Cowpeas": "•Cowpeas•",
   "cowpeas": "•cowpeas•",
   "Create Calendar": "•Create Calendar•",
-  "Create calender": "•Create calender•",
   "Create New": "•Create New•",
   "Create New Budget": "Unda bajeti mpya",
   "Crop": "•Crop•",
@@ -139,6 +141,7 @@
   "Hide Options": "•Hide Options•",
   "High": "•High•",
   "Historic climate": "•Historic climate•",
+  "hours": "•hours•",
   "housing": "•housing•",
   "How long is your budget for?": "•How long is your budget for?•",
   "Import": "•Import•",
@@ -229,7 +232,7 @@
   "Pigeon Peas": "•Pigeon Peas•",
   "pigeon peas": "•pigeon peas•",
   "pigs": "•pigs•",
-  "Please add the calender title": "•Please add the calender title•",
+  "Please add title": "•Please add title•",
   "Please enter start month": "•Please enter start month•",
   "Please enter total months": "•Please enter total months•",
   "Please select at least 1 crop": "•Please select at least 1 crop•",
