+++ libs/i18n/assets/tg.json
@@ -29,6 +29,7 @@
   "Bad": "•Bad•",
   "bagging": "халта",
   "bags": "халта",
+  "balance": "•balance•",
   "Balance": "Мувозинат",
   "banding": "банд",
   "Beans": "•Beans•",
@@ -42,8 +43,8 @@
   "build housing": "сохтани манзил",
   "Calculated by using FAO CLIMWAT 2.0 for Cropwat and Cropwat 8.0 and climate data for Nkhotakota Met Station.": "•Calculated by using FAO CLIMWAT 2.0 for Cropwat and Cropwat 8.0 and climate data for Nkhotakota Met Station.•",
   "Calculated with the date range of Start of rains and End of the rains": "•Calculated with the date range of Start of rains and End of the rains•",
+  "Calendar name": "•Calendar name•",
   "Calendar Time": "•Calendar Time•",
-  "Calender name": "•Calender name•",
   "Cancel": "Манъ кардан",
   "Capacity of 100mm on 1st March reduces 0.5 with evaporation rate of 5mm per day": "•Capacity of 100mm on 1st March reduces 0.5 with evaporation rate of 5mm per day•",
   "Cassava": "•Cassava•",
@@ -62,13 +63,14 @@
   "Compare different options and plans": "•Compare different options and plans•",
   "Comparing different options and planning using Participatory Budgets": "•Comparing different options and planning using Participatory Budgets•",
   "compost manure making": "компост компост",
+  "consumed": "•consumed•",
   "Consumed": "•Consumed•",
+  "Copy": "•Copy•",
   "Cost": "•Cost•",
   "Cotton": "•Cotton•",
   "Cowpeas": "•Cowpeas•",
   "cowpeas": "гов",
   "Create Calendar": "•Create Calendar•",
-  "Create calender": "•Create calender•",
   "Create New": "•Create New•",
   "Create New Budget": "Эҷоди буҷети нав",
   "Crop": "•Crop•",
@@ -139,6 +141,7 @@
   "Hide Options": "•Hide Options•",
   "High": "•High•",
   "Historic climate": "•Historic climate•",
+  "hours": "•hours•",
   "housing": "манзил",
   "How long is your budget for?": "Буҷаи шумо чанд вақт аст?",
   "Import": "•Import•",
@@ -229,7 +232,7 @@
   "Pigeon Peas": "•Pigeon Peas•",
   "pigeon peas": "нахуд кабӯтар",
   "pigs": "хукҳо",
-  "Please add the calender title": "•Please add the calender title•",
+  "Please add title": "•Please add title•",
   "Please enter start month": "•Please enter start month•",
   "Please enter total months": "•Please enter total months•",
   "Please select at least 1 crop": "•Please select at least 1 crop•",
