+++ libs/i18n/assets/zm_ny.json
@@ -29,6 +29,7 @@
   "Bad": "•Bad•",
   "bagging": "kuika m'mathumba",
   "bags": "matumba",
+  "balance": "•balance•",
   "Balance": "Zokharira",
   "banding": "kumangilira",
   "Beans": "Kaela",
@@ -42,8 +43,8 @@
   "build housing": "kumanga nyumba",
   "Calculated by using FAO CLIMWAT 2.0 for Cropwat and Cropwat 8.0 and climate data for Nkhotakota Met Station.": "Kuwerengela kusewenzetsa FAO CLIMWAT 2.0 for Cropwat and Cropwat 8.0 and climate data for Nkhotakota Met Station.",
   "Calculated with the date range of Start of rains and End of the rains": "•Calculated with the date range of Start of rains and End of the rains•",
+  "Calendar name": "•Calendar name•",
   "Calendar Time": "•Calendar Time•",
-  "Calender name": "•Calender name•",
   "Cancel": "kutseka",
   "Capacity of 100mm on 1st March reduces 0.5 with evaporation rate of 5mm per day": "•Capacity of 100mm on 1st March reduces 0.5 with evaporation rate of 5mm per day•",
   "Cassava": "Chinangwa",
@@ -62,13 +63,14 @@
   "Compare different options and plans": "Linganizani Zosankha",
   "Comparing different options and planning using Participatory Budgets": "•Comparing different options and planning using Participatory Budgets•",
   "compost manure making": "kupanga manyowa kuzinyalala",
+  "consumed": "•consumed•",
   "Consumed": "Zosewenzetsedwa",
+  "Copy": "•Copy•",
   "Cost": "Mtengo",
   "Cotton": "•Cotton•",
   "Cowpeas": "•Cowpeas•",
   "cowpeas": "nyemba",
   "Create Calendar": "•Create Calendar•",
-  "Create calender": "•Create calender•",
   "Create New": "•Create New•",
   "Create New Budget": "Pangani bajeti yatsopano",
   "Crop": "•Crop•",
@@ -139,6 +141,7 @@
   "Hide Options": "Bimphani zosankha",
   "High": "•High•",
   "Historic climate": "Mbiri ya za Nyengo",
+  "hours": "•hours•",
   "housing": "mosungila",
   "How long is your budget for?": "bajeti yanu imakhala yayitali motani?",
   "Import": "Kuika nji",
@@ -229,7 +232,7 @@
   "Pigeon Peas": "•Pigeon Peas•",
   "pigeon peas": "nyamdolo",
   "pigs": "nkhumba",
-  "Please add the calender title": "•Please add the calender title•",
+  "Please add title": "•Please add title•",
   "Please enter start month": "•Please enter start month•",
   "Please enter total months": "•Please enter total months•",
   "Please select at least 1 crop": "•Please select at least 1 crop•",
