+++ libs/i18n/templates/_template.csv
@@ -118,8 +118,11 @@ tool,context,text
 "budget",,"Total Consumed"
 "budget",,"Total Produced"
 "budget",,"What is your type of enterprise?"
+"budget",,"balance"
+"budget",,"consumed"
 "budget",,"description"
 "budget",,"fruits"
+"budget",,"hours"
 "budget",,"livestock"
 "budget",,"starting in"
 "budget",,"title"
@@ -372,13 +375,13 @@ tool,context,text
 "resources",,"Title"
 "resources",,"Type"
 "seasonal-calendar",,"Calendar Time"
-"seasonal-calendar",,"Calender name"
+"seasonal-calendar",,"Calendar name"
+"seasonal-calendar",,"Copy"
 "seasonal-calendar",,"Create Calendar"
-"seasonal-calendar",,"Create calender"
 "seasonal-calendar",,"Example Calendar"
 "seasonal-calendar",,"No saved calendars"
 "seasonal-calendar",,"Number of Months"
-"seasonal-calendar",,"Please add the calender title"
+"seasonal-calendar",,"Please add title"
 "seasonal-calendar",,"Please enter start month"
 "seasonal-calendar",,"Please enter total months"
 "seasonal-calendar",,"Please select at least 1 crop"
